feat: clean code architecture conversion of user management module #1125
