--- admin-ui/plugins/user-management/common/Constants.js
@@ -1,2 +1,0 @@
-export const LIMIT_ID = 'searchLimit'
-export const PATTERN_ID = 'searchPattern'
