--- admin-ui/plugins/user-management/common/countries.js
@@ -1,1002 +1,0 @@
-export const countries = [
-  {
-    cca2: 'DE',
-    name: 'Germany',
-  },
-  {
-    cca2: 'MA',
-    name: 'Morocco',
-  },
-  {
-    cca2: 'BE',
-    name: 'Belgium',
-  },
-  {
-    cca2: 'CK',
-    name: 'Cook Islands',
-  },
-  {
-    cca2: 'EC',
-    name: 'Ecuador',
-  },
-  {
-    cca2: 'KN',
-    name: 'Saint Kitts and Nevis',
-  },
-  {
-    cca2: 'RW',
-    name: 'Rwanda',
-  },
-  {
-    cca2: 'GY',
-    name: 'Guyana',
-  },
-  {
-    cca2: 'MP',
-    name: 'Northern Mariana Islands',
-  },
-  {
-    cca2: 'BW',
-    name: 'Botswana',
-  },
-  {
-    cca2: 'BV',
-    name: 'Bouvet Island',
-  },
-  {
-    cca2: 'CN',
-    name: 'China',
-  },
-  {
-    cca2: 'GD',
-    name: 'Grenada',
-  },
-  {
-    cca2: 'SN',
-    name: 'Senegal',
-  },
-  {
-    cca2: 'GU',
-    name: 'Guam',
-  },
-  {
-    cca2: 'CG',
-    name: 'Republic of the Congo',
-  },
-  {
-    cca2: 'TH',
-    name: 'Thailand',
-  },
-  {
-    cca2: 'JO',
-    name: 'Jordan',
-  },
-  {
-    cca2: 'CO',
-    name: 'Colombia',
-  },
-  {
-    cca2: 'PN',
-    name: 'Pitcairn Islands',
-  },
-  {
-    cca2: 'NO',
-    name: 'Norway',
-  },
-  {
-    cca2: 'KM',
-    name: 'Comoros',
-  },
-  {
-    cca2: 'TF',
-    name: 'French Southern and Antarctic Lands',
-  },
-  {
-    cca2: 'CV',
-    name: 'Cape Verde',
-  },
-  {
-    cca2: 'CY',
-    name: 'Cyprus',
-  },
-  {
-    cca2: 'KI',
-    name: 'Kiribati',
-  },
-  {
-    cca2: 'CF',
-    name: 'Central African Republic',
-  },
-  {
-    cca2: 'BQ',
-    name: 'Caribbean Netherlands',
-  },
-  {
-    cca2: 'PR',
-    name: 'Puerto Rico',
-  },
-  {
-    cca2: 'CC',
-    name: 'Cocos (Keeling) Islands',
-  },
-  {
-    cca2: 'PS',
-    name: 'Palestine',
-  },
-  {
-    cca2: 'TM',
-    name: 'Turkmenistan',
-  },
-  {
-    cca2: 'IO',
-    name: 'British Indian Ocean Territory',
-  },
-  {
-    cca2: 'QA',
-    name: 'Qatar',
-  },
-  {
-    cca2: 'GL',
-    name: 'Greenland',
-  },
-  {
-    cca2: 'RO',
-    name: 'Romania',
-  },
-  {
-    cca2: 'SH',
-    name: 'Saint Helena, Ascension and Tristan da Cunha',
-  },
-  {
-    cca2: 'NU',
-    name: 'Niue',
-  },
-  {
-    cca2: 'IN',
-    name: 'India',
-  },
-  {
-    cca2: 'RE',
-    name: 'Réunion',
-  },
-  {
-    cca2: 'UZ',
-    name: 'Uzbekistan',
-  },
-  {
-    cca2: 'AM',
-    name: 'Armenia',
-  },
-  {
-    cca2: 'KW',
-    name: 'Kuwait',
-  },
-  {
-    cca2: 'SX',
-    name: 'Sint Maarten',
-  },
-  {
-    cca2: 'BG',
-    name: 'Bulgaria',
-  },
-  {
-    cca2: 'IL',
-    name: 'Israel',
-  },
-  {
-    cca2: 'UY',
-    name: 'Uruguay',
-  },
-  {
-    cca2: 'VU',
-    name: 'Vanuatu',
-  },
-  {
-    cca2: 'VN',
-    name: 'Vietnam',
-  },
-  {
-    cca2: 'AS',
-    name: 'American Samoa',
-  },
-  {
-    cca2: 'BF',
-    name: 'Burkina Faso',
-  },
-  {
-    cca2: 'PA',
-    name: 'Panama',
-  },
-  {
-    cca2: 'GF',
-    name: 'French Guiana',
-  },
-  {
-    cca2: 'IQ',
-    name: 'Iraq',
-  },
-  {
-    cca2: 'TW',
-    name: 'Taiwan',
-  },
-  {
-    cca2: 'CI',
-    name: 'Ivory Coast',
-  },
-  {
-    cca2: 'LK',
-    name: 'Sri Lanka',
-  },
-  {
-    cca2: 'MD',
-    name: 'Moldova',
-  },
-  {
-    cca2: 'GB',
-    name: 'United Kingdom',
-  },
-  {
-    cca2: 'GR',
-    name: 'Greece',
-  },
-  {
-    cca2: 'ST',
-    name: 'São Tomé and Príncipe',
-  },
-  {
-    cca2: 'ML',
-    name: 'Mali',
-  },
-  {
-    cca2: 'ZA',
-    name: 'South Africa',
-  },
-  {
-    cca2: 'IS',
-    name: 'Iceland',
-  },
-  {
-    cca2: 'KR',
-    name: 'South Korea',
-  },
-  {
-    cca2: 'CL',
-    name: 'Chile',
-  },
-  {
-    cca2: 'MU',
-    name: 'Mauritius',
-  },
-  {
-    cca2: 'XK',
-    name: 'Kosovo',
-  },
-  {
-    cca2: 'SI',
-    name: 'Slovenia',
-  },
-  {
-    cca2: 'CZ',
-    name: 'Czechia',
-  },
-  {
-    cca2: 'SM',
-    name: 'San Marino',
-  },
-  {
-    cca2: 'BY',
-    name: 'Belarus',
-  },
-  {
-    cca2: 'GE',
-    name: 'Georgia',
-  },
-  {
-    cca2: 'CH',
-    name: 'Switzerland',
-  },
-  {
-    cca2: 'CU',
-    name: 'Cuba',
-  },
-  {
-    cca2: 'NF',
-    name: 'Norfolk Island',
-  },
-  {
-    cca2: 'CD',
-    name: 'DR Congo',
-  },
-  {
-    cca2: 'MF',
-    name: 'Saint Martin',
-  },
-  {
-    cca2: 'SE',
-    name: 'Sweden',
-  },
-  {
-    cca2: 'HU',
-    name: 'Hungary',
-  },
-  {
-    cca2: 'AW',
-    name: 'Aruba',
-  },
-  {
-    cca2: 'AE',
-    name: 'United Arab Emirates',
-  },
-  {
-    cca2: 'ME',
-    name: 'Montenegro',
-  },
-  {
-    cca2: 'KP',
-    name: 'North Korea',
-  },
-  {
-    cca2: 'TJ',
-    name: 'Tajikistan',
-  },
-  {
-    cca2: 'LY',
-    name: 'Libya',
-  },
-  {
-    cca2: 'YE',
-    name: 'Yemen',
-  },
-  {
-    cca2: 'NL',
-    name: 'Netherlands',
-  },
-  {
-    cca2: 'HM',
-    name: 'Heard Island and McDonald Islands',
-  },
-  {
-    cca2: 'NC',
-    name: 'New Caledonia',
-  },
-  {
-    cca2: 'KZ',
-    name: 'Kazakhstan',
-  },
-  {
-    cca2: 'VG',
-    name: 'British Virgin Islands',
-  },
-  {
-    cca2: 'LB',
-    name: 'Lebanon',
-  },
-  {
-    cca2: 'NG',
-    name: 'Nigeria',
-  },
-  {
-    cca2: 'AG',
-    name: 'Antigua and Barbuda',
-  },
-  {
-    cca2: 'PL',
-    name: 'Poland',
-  },
-  {
-    cca2: 'BA',
-    name: 'Bosnia and Herzegovina',
-  },
-  {
-    cca2: 'AU',
-    name: 'Australia',
-  },
-  {
-    cca2: 'DZ',
-    name: 'Algeria',
-  },
-  {
-    cca2: 'MG',
-    name: 'Madagascar',
-  },
-  {
-    cca2: 'TT',
-    name: 'Trinidad and Tobago',
-  },
-  {
-    cca2: 'SR',
-    name: 'Suriname',
-  },
-  {
-    cca2: 'NP',
-    name: 'Nepal',
-  },
-  {
-    cca2: 'ID',
-    name: 'Indonesia',
-  },
-  {
-    cca2: 'AZ',
-    name: 'Azerbaijan',
-  },
-  {
-    cca2: 'ES',
-    name: 'Spain',
-  },
-  {
-    cca2: 'SV',
-    name: 'El Salvador',
-  },
-  {
-    cca2: 'HK',
-    name: 'Hong Kong',
-  },
-  {
-    cca2: 'TK',
-    name: 'Tokelau',
-  },
-  {
-    cca2: 'MH',
-    name: 'Marshall Islands',
-  },
-  {
-    cca2: 'RU',
-    name: 'Russia',
-  },
-  {
-    cca2: 'LC',
-    name: 'Saint Lucia',
-  },
-  {
-    cca2: 'TV',
-    name: 'Tuvalu',
-  },
-  {
-    cca2: 'GQ',
-    name: 'Equatorial Guinea',
-  },
-  {
-    cca2: 'MY',
-    name: 'Malaysia',
-  },
-  {
-    cca2: 'MW',
-    name: 'Malawi',
-  },
-  {
-    cca2: 'SC',
-    name: 'Seychelles',
-  },
-  {
-    cca2: 'MT',
-    name: 'Malta',
-  },
-  {
-    cca2: 'DM',
-    name: 'Dominica',
-  },
-  {
-    cca2: 'OM',
-    name: 'Oman',
-  },
-  {
-    cca2: 'AO',
-    name: 'Angola',
-  },
-  {
-    cca2: 'PT',
-    name: 'Portugal',
-  },
-  {
-    cca2: 'CM',
-    name: 'Cameroon',
-  },
-  {
-    cca2: 'EE',
-    name: 'Estonia',
-  },
-  {
-    cca2: 'ER',
-    name: 'Eritrea',
-  },
-  {
-    cca2: 'MR',
-    name: 'Mauritania',
-  },
-  {
-    cca2: 'MS',
-    name: 'Montserrat',
-  },
-  {
-    cca2: 'CA',
-    name: 'Canada',
-  },
-  {
-    cca2: 'LU',
-    name: 'Luxembourg',
-  },
-  {
-    cca2: 'AL',
-    name: 'Albania',
-  },
-  {
-    cca2: 'PM',
-    name: 'Saint Pierre and Miquelon',
-  },
-  {
-    cca2: 'YT',
-    name: 'Mayotte',
-  },
-  {
-    cca2: 'FJ',
-    name: 'Fiji',
-  },
-  {
-    cca2: 'SL',
-    name: 'Sierra Leone',
-  },
-  {
-    cca2: 'TC',
-    name: 'Turks and Caicos Islands',
-  },
-  {
-    cca2: 'CW',
-    name: 'Curaçao',
-  },
-  {
-    cca2: 'UA',
-    name: 'Ukraine',
-  },
-  {
-    cca2: 'MK',
-    name: 'North Macedonia',
-  },
-  {
-    cca2: 'AD',
-    name: 'Andorra',
-  },
-  {
-    cca2: 'NA',
-    name: 'Namibia',
-  },
-  {
-    cca2: 'WS',
-    name: 'Samoa',
-  },
-  {
-    cca2: 'SY',
-    name: 'Syria',
-  },
-  {
-    cca2: 'AR',
-    name: 'Argentina',
-  },
-  {
-    cca2: 'EH',
-    name: 'Western Sahara',
-  },
-  {
-    cca2: 'MZ',
-    name: 'Mozambique',
-  },
-  {
-    cca2: 'GN',
-    name: 'Guinea',
-  },
-  {
-    cca2: 'TN',
-    name: 'Tunisia',
-  },
-  {
-    cca2: 'FK',
-    name: 'Falkland Islands',
-  },
-  {
-    cca2: 'LS',
-    name: 'Lesotho',
-  },
-  {
-    cca2: 'HR',
-    name: 'Croatia',
-  },
-  {
-    cca2: 'LR',
-    name: 'Liberia',
-  },
-  {
-    cca2: 'PG',
-    name: 'Papua New Guinea',
-  },
-  {
-    cca2: 'FI',
-    name: 'Finland',
-  },
-  {
-    cca2: 'AQ',
-    name: 'Antarctica',
-  },
-  {
-    cca2: 'BI',
-    name: 'Burundi',
-  },
-  {
-    cca2: 'BO',
-    name: 'Bolivia',
-  },
-  {
-    cca2: 'TG',
-    name: 'Togo',
-  },
-  {
-    cca2: 'BB',
-    name: 'Barbados',
-  },
-  {
-    cca2: 'GI',
-    name: 'Gibraltar',
-  },
-  {
-    cca2: 'ET',
-    name: 'Ethiopia',
-  },
-  {
-    cca2: 'TD',
-    name: 'Chad',
-  },
-  {
-    cca2: 'GM',
-    name: 'Gambia',
-  },
-  {
-    cca2: 'NZ',
-    name: 'New Zealand',
-  },
-  {
-    cca2: 'UM',
-    name: 'United States Minor Outlying Islands',
-  },
-  {
-    cca2: 'TL',
-    name: 'Timor-Leste',
-  },
-  {
-    cca2: 'TR',
-    name: 'Turkey',
-  },
-  {
-    cca2: 'KH',
-    name: 'Cambodia',
-  },
-  {
-    cca2: 'LT',
-    name: 'Lithuania',
-  },
-  {
-    cca2: 'SD',
-    name: 'Sudan',
-  },
-  {
-    cca2: 'AI',
-    name: 'Anguilla',
-  },
-  {
-    cca2: 'IT',
-    name: 'Italy',
-  },
-  {
-    cca2: 'MV',
-    name: 'Maldives',
-  },
-  {
-    cca2: 'PF',
-    name: 'French Polynesia',
-  },
-  {
-    cca2: 'AT',
-    name: 'Austria',
-  },
-  {
-    cca2: 'BZ',
-    name: 'Belize',
-  },
-  {
-    cca2: 'KG',
-    name: 'Kyrgyzstan',
-  },
-  {
-    cca2: 'LA',
-    name: 'Laos',
-  },
-  {
-    cca2: 'DO',
-    name: 'Dominican Republic',
-  },
-  {
-    cca2: 'HN',
-    name: 'Honduras',
-  },
-  {
-    cca2: 'SZ',
-    name: 'Eswatini',
-  },
-  {
-    cca2: 'GW',
-    name: 'Guinea-Bissau',
-  },
-  {
-    cca2: 'NE',
-    name: 'Niger',
-  },
-  {
-    cca2: 'PE',
-    name: 'Peru',
-  },
-  {
-    cca2: 'BH',
-    name: 'Bahrain',
-  },
-  {
-    cca2: 'BL',
-    name: 'Saint Barthélemy',
-  },
-  {
-    cca2: 'BN',
-    name: 'Brunei',
-  },
-  {
-    cca2: 'KY',
-    name: 'Cayman Islands',
-  },
-  {
-    cca2: 'PY',
-    name: 'Paraguay',
-  },
-  {
-    cca2: 'BJ',
-    name: 'Benin',
-  },
-  {
-    cca2: 'IR',
-    name: 'Iran',
-  },
-  {
-    cca2: 'MO',
-    name: 'Macau',
-  },
-  {
-    cca2: 'MX',
-    name: 'Mexico',
-  },
-  {
-    cca2: 'JE',
-    name: 'Jersey',
-  },
-  {
-    cca2: 'SO',
-    name: 'Somalia',
-  },
-  {
-    cca2: 'LV',
-    name: 'Latvia',
-  },
-  {
-    cca2: 'TZ',
-    name: 'Tanzania',
-  },
-  {
-    cca2: 'VE',
-    name: 'Venezuela',
-  },
-  {
-    cca2: 'MN',
-    name: 'Mongolia',
-  },
-  {
-    cca2: 'CR',
-    name: 'Costa Rica',
-  },
-  {
-    cca2: 'TO',
-    name: 'Tonga',
-  },
-  {
-    cca2: 'BR',
-    name: 'Brazil',
-  },
-  {
-    cca2: 'ZW',
-    name: 'Zimbabwe',
-  },
-  {
-    cca2: 'BT',
-    name: 'Bhutan',
-  },
-  {
-    cca2: 'KE',
-    name: 'Kenya',
-  },
-  {
-    cca2: 'VC',
-    name: 'Saint Vincent and the Grenadines',
-  },
-  {
-    cca2: 'AX',
-    name: 'Åland Islands',
-  },
-  {
-    cca2: 'IM',
-    name: 'Isle of Man',
-  },
-  {
-    cca2: 'IE',
-    name: 'Ireland',
-  },
-  {
-    cca2: 'HT',
-    name: 'Haiti',
-  },
-  {
-    cca2: 'MM',
-    name: 'Myanmar',
-  },
-  {
-    cca2: 'GP',
-    name: 'Guadeloupe',
-  },
-  {
-    cca2: 'JM',
-    name: 'Jamaica',
-  },
-  {
-    cca2: 'BS',
-    name: 'Bahamas',
-  },
-  {
-    cca2: 'FR',
-    name: 'France',
-  },
-  {
-    cca2: 'SG',
-    name: 'Singapore',
-  },
-  {
-    cca2: 'WF',
-    name: 'Wallis and Futuna',
-  },
-  {
-    cca2: 'FO',
-    name: 'Faroe Islands',
-  },
-  {
-    cca2: 'AF',
-    name: 'Afghanistan',
-  },
-  {
-    cca2: 'SJ',
-    name: 'Svalbard and Jan Mayen',
-  },
-  {
-    cca2: 'SS',
-    name: 'South Sudan',
-  },
-  {
-    cca2: 'RS',
-    name: 'Serbia',
-  },
-  {
-    cca2: 'VA',
-    name: 'Vatican City',
-  },
-  {
-    cca2: 'NI',
-    name: 'Nicaragua',
-  },
-  {
-    cca2: 'UG',
-    name: 'Uganda',
-  },
-  {
-    cca2: 'US',
-    name: 'United States',
-  },
-  {
-    cca2: 'EG',
-    name: 'Egypt',
-  },
-  {
-    cca2: 'MQ',
-    name: 'Martinique',
-  },
-  {
-    cca2: 'PH',
-    name: 'Philippines',
-  },
-  {
-    cca2: 'GS',
-    name: 'South Georgia',
-  },
-  {
-    cca2: 'ZM',
-    name: 'Zambia',
-  },
-  {
-    cca2: 'SB',
-    name: 'Solomon Islands',
-  },
-  {
-    cca2: 'GH',
-    name: 'Ghana',
-  },
-  {
-    cca2: 'GA',
-    name: 'Gabon',
-  },
-  {
-    cca2: 'PK',
-    name: 'Pakistan',
-  },
-  {
-    cca2: 'SA',
-    name: 'Saudi Arabia',
-  },
-  {
-    cca2: 'GG',
-    name: 'Guernsey',
-  },
-  {
-    cca2: 'GT',
-    name: 'Guatemala',
-  },
-  {
-    cca2: 'NR',
-    name: 'Nauru',
-  },
-  {
-    cca2: 'MC',
-    name: 'Monaco',
-  },
-  {
-    cca2: 'CX',
-    name: 'Christmas Island',
-  },
-  {
-    cca2: 'JP',
-    name: 'Japan',
-  },
-  {
-    cca2: 'LI',
-    name: 'Liechtenstein',
-  },
-  {
-    cca2: 'PW',
-    name: 'Palau',
-  },
-  {
-    cca2: 'DJ',
-    name: 'Djibouti',
-  },
-  {
-    cca2: 'VI',
-    name: 'United States Virgin Islands',
-  },
-  {
-    cca2: 'BM',
-    name: 'Bermuda',
-  },
-  {
-    cca2: 'SK',
-    name: 'Slovakia',
-  },
-  {
-    cca2: 'BD',
-    name: 'Bangladesh',
-  },
-  {
-    cca2: 'DK',
-    name: 'Denmark',
-  },
-  {
-    cca2: 'FM',
-    name: 'Micronesia',
-  },
-]
