+++ admin-ui/plugins/user-management/domain/entities/TUserInputs.ts
@@ -1,0 +1,18 @@
+export interface TDefaultUserInputs {
+  userId: string;
+  mail: string;
+  displayName: string;
+  jansStatus: string;
+  givenName: string;
+  userPassword?: string;
+  customAttributes?: Array<TCustomAttributes>;
+  sn?: string;
+  middleName?: string;
+  userConfirmPassword?: string;
+}
+
+export interface TCustomAttributes {
+  name: string;
+  multiValued: boolean;
+  values: Array<string>;
+}
