+++ admin-ui/plugins/user-management/domain/redux/audit/Resources.js
