+++ admin-ui/plugins/user-management/presentation/common/Constants.js
@@ -1,0 +1,2 @@
+export const LIMIT_ID = 'searchLimit'
+export const PATTERN_ID = 'searchPattern'
